extracted renderLink function for cleaner NavBar
