+++ components/Nav/NavBar.tsx
@@ -4,41 +4,13 @@ import Link from "next/link"
 import Image from "next/image";
 import data from "@/lib/data/pages.json";
 import { useState } from "react";
-import SearchBar from "./search-bar";
-import { usePathname } from "next/navigation";
+import SearchBar from "../search-bar";
+import { RenderLink } from "./render-link";
 
-type PageLink = {
-    href: string;
-    label: string;
-};
 
 export default function NavBar() {
     const [menuOpen, setMenuOpen] = useState(false);
     const [searchOpen, setSearchOpen] = useState(false);
-    const pathname = usePathname();
-
-    // ids for anchor links on start page
-    const anchorLinks = {
-        Home: "#home",
-        About: "#about",
-        Contact: "#contact",
-    };
-
-    const renderLink = (link: PageLink) => {
-        // if on start page and link is Home, About or Contact, render anchor link
-        if (pathname === "/" &&
-            (link.label === "Home" || link.label === "About" || link.label === "Contact")
-        ) {
-            return (
-                <Link href={anchorLinks[link.label]}>{link.label}</Link>
-            );
-        } else {
-            // otherwise normal navigation link
-            return (
-                <Link href={link.href}>{link.label}</Link>
-            );
-        }
-    };
 
     return (
         <div className="relative">
@@ -49,7 +21,9 @@ export default function NavBar() {
                 {/* desktop nav */}
                 <ul className="hidden md:flex gap-6">
                     {data["pages"].map((link, index) => (
-                    <li key={index} className="text-neutral-800 hover:text-blue-900">{renderLink(link)}</li>
+                    <li key={index} className="text-neutral-800 hover:text-blue-900">
+                        <RenderLink {...link} />
+                    </li>
                     ))}
                 </ul>
             </div>
@@ -58,7 +32,8 @@ export default function NavBar() {
                   <div className="md:flex text-1xl font-semibold md:justify-start md:w-1/3">
                     <ul className="flex flex-col absolute top-20 left-0 w-full px-4 bg-white text-neutral-800 md:flex-row md:gap-8 md:static md:top-auto md:left-auto md:font-normal">
                         {data["pages"].map((link, index) => (
-                        <li key={index} className="flex items-center justify-between py-2 text-neutral-800 hover:text-blue-900">{renderLink(link)}
+                        <li key={index} className="flex items-center justify-between py-2 text-neutral-800 hover:text-blue-900">
+                            <RenderLink {...link} />
                             <svg xmlns="http://www.w3.org/2000/svg" height="16px" viewBox="0 -960 960 960" width="16px" fill="currentColor" className="w-3 h-3 md:hidden">
                                 <path d="m321-80-71-71 329-329-329-329 71-71 400 400L321-80Z"/>
                             </svg>
@@ -74,6 +49,7 @@ export default function NavBar() {
                     <h2 className="text-3xl font-bold text-blue-900">MegaMart</h2>
                 </Link>
             </div >
+
             {/* icons */}
             <div className="flex md:w-1/3 md:justify-end">
                 {/* link icons + hamburger menu button */}
